requirements.txt
appdirs==1.4.0
bitaddress==0.1
cryptonator==0.0.2
Django==1.10.5
django-countries==4.0
django-sendfile==0.3.11
ecdsa==0.13
gevent==1.2.1
greenlet==0.4.11
packaging==16.8
pluggy==0.4.0
py==1.4.32
pyjsonrpc==0.0.2
pyotp==2.2.4
pyparsing==2.1.10
python-bitcoinrpc==1.0
qrcode==5.3
requests==2.13.0
six==1.10.0
tox==2.5.0
virtualenv==15.1.0
ws4py==0.3.5
